requirements.txt
# Core LlamaIndex
llama-index>=0.11.0
llama-index-graph-stores-neo4j>=0.3.0
llama-index-llms-openai>=0.2.0
llama-index-embeddings-openai>=0.2.0

# Neo4j Driver (for direct graph queries)
neo4j>=5.0.0

# Web UI
streamlit>=1.28.0

# Graph Visualization
pyvis>=0.3.0

# Configuration
python-dotenv>=1.0.0
